requirements.txt
annotated-types==0.7.0
anyio==4.12.1
audioread==3.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
cuda-bindings==12.9.4
cuda-pathfinder==1.4.2
decorator==5.2.1
distro==1.9.0
einops==0.8.2
exceptiongroup==1.3.1
filelock==3.25.2
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
ImageIO==2.37.3
imageio-ffmpeg==0.6.0
inflect==7.5.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
lazy-loader==0.5
librosa==0.11.0
llvmlite==0.46.0
MarkupSafe==3.0.3
more-itertools==10.8.0
moviepy==2.2.1
mpmath==1.3.0
msgpack==1.1.2
networkx==3.4.2
numba==0.64.0
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
openai==2.26.0
packaging==26.0
pillow==11.3.0
platformdirs==4.9.4
pooch==1.9.0
proglog==0.1.12
progressbar==2.5
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
rotary-embedding-torch==0.8.9
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.13.3
torch==2.10.0
torchaudio==2.10.0
torchcodec==0.10.0
tortoise-tts==3.0.0
tqdm==4.67.3
transformers==4.31.0
triton==3.6.0
typeguard==4.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
Unidecode==1.4.0
urllib3==2.6.3
